requirements.txt
awsiotsdk>=1.18.0
boto3>=1.34.0
botocore>=1.34.0
